Split matchup card into multiple components. Updated view models
