+++ src/services/ApiService.ts
@@ -1,5 +1,5 @@
 import http from "../http-common";
-import IGameOdds from "../types/GameOdds";
+import {IGameOdds} from "../types/GameOdds";
 
 const getAllPredictedGames = () => {
   return http.get<Array<IGameOdds>>("/PredictedGame/GetAllPredictedGames/");
